Birthday sort order in box
Birthdays are currently sorted by date of birth and not by the day in a month of the day of birth. This leads to names appearing in the "wrong" order. Example:

* Person A is born on 1980-01-10
* Person B is born on 1970-01-30

When I see the list of birthdays for January, I expect Person A to show up before Person B, but without this patch, the opposite is the case.